requirements.txt
Flask==0.12.1
PyYAML==3.12
